Switch to a makefile system
Some side effects: Once we get going, spinup seems slightly better?!

However, the motor gets extremely hot very fast, we had the magic smoke escape almost instantly! Not sure how to combat that one without letting the ESC brown out constantly...

However, with the power very very low, and a admittedly rocky start (that might be due to the magic smoke being released, or this being a bad ESC to begin with).

RE: This being a bad ESC to begin with.

This ESC has historically given bad starts, it was in my bin labelled BAD in sharpie, which means if I recall correctly, that

- It had resistor network problems and sometimes had the classic start delay, which shouldn't be a problem.
But it might also be something worse.

We might be getting farenough along that it's time to ditch the `bad esc` and motor pair and use hardware I care about/have tested.

But first, before I think about that....
- PRoperly split into .cc/.h files
- Remove all uses of byte/short, replace with uint*.
- Get my naming consistent
- Remove the enum, replace with pointer.
- Review the diff of this commit in more detail when I am not sleep deprived at 1 am.